--- worker/pkg/workflows/datasync/activities/gen-benthos-configs/mysql.go
@@ -1,34 +1,0 @@
-package genbenthosconfigs_activity
-
-import (
-	"github.com/xwb1989/sqlparser"
-)
-
-func qualifyMysqlWhereColumnNames(sql, schema, table string) (string, error) {
-	stmt, err := sqlparser.Parse(sql)
-	if err != nil {
-		return "", err
-	}
-
-	switch stmt := stmt.(type) { //nolint:gocritic
-	case *sqlparser.Select:
-		err = sqlparser.Walk(func(node sqlparser.SQLNode) (kontinue bool, err error) {
-			switch node := node.(type) { //nolint:gocritic
-			case *sqlparser.ComparisonExpr:
-				if col, ok := node.Left.(*sqlparser.ColName); ok {
-					if col.Qualifier.IsEmpty() {
-						col.Qualifier.Qualifier = sqlparser.NewTableIdent(schema)
-						col.Qualifier.Name = sqlparser.NewTableIdent(table)
-					}
-				}
-				return false, nil
-			}
-			return true, nil
-		}, stmt)
-		if err != nil {
-			return "", err
-		}
-	}
-
-	return sqlparser.String(stmt), nil
-}
