+++ worker/pkg/workflows/datasync/activities/gen-benthos-configs/query-builder.go
@@ -11,6 +11,9 @@ import (
 	sql_manager "github.com/nucleuscloud/neosync/backend/pkg/sqlmanager"
 	tabledependency "github.com/nucleuscloud/neosync/backend/pkg/table-dependency"
 	"github.com/nucleuscloud/neosync/worker/pkg/workflows/datasync/activities/shared"
+	pg_query "github.com/pganalyze/pg_query_go/v5"
+	pgquery "github.com/wasilibs/go-pgquery"
+	"github.com/xwb1989/sqlparser"
 
 	// import the dialect
 	_ "github.com/doug-martin/goqu/v9/dialect/mysql"
@@ -592,4 +595,91 @@ func reverseSlice[T any](slice []T) {
 	for i, j := 0, len(slice)-1; i < j; i, j = i+1, j-1 {
 		slice[i], slice[j] = slice[j], slice[i]
 	}
+}
+
+func qualifyPostgresWhereColumnNames(sql, schema, table string) (string, error) {
+	tree, err := pgquery.Parse(sql)
+	if err != nil {
+		return "", err
+	}
+
+	for _, stmt := range tree.GetStmts() {
+		selectStmt := stmt.GetStmt().GetSelectStmt()
+
+		if selectStmt.WhereClause != nil {
+			updatePostgresExpr(schema, table, selectStmt.WhereClause)
+		}
+	}
+	updatedSql, err := pgquery.Deparse(tree)
+	if err != nil {
+		return "", err
+	}
+	return updatedSql, nil
+}
+
+func updatePostgresExpr(schema, table string, node *pg_query.Node) {
+	switch expr := node.Node.(type) {
+	case *pg_query.Node_SubLink:
+		updatePostgresExpr(schema, table, node.GetSubLink().GetTestexpr())
+	case *pg_query.Node_BoolExpr:
+		for _, arg := range expr.BoolExpr.GetArgs() {
+			updatePostgresExpr(schema, table, arg)
+		}
+	case *pg_query.Node_AExpr:
+		updatePostgresExpr(schema, table, expr.AExpr.GetLexpr())
+		updatePostgresExpr(schema, table, expr.AExpr.Rexpr)
+	case *pg_query.Node_ColumnDef:
+	case *pg_query.Node_ColumnRef:
+		col := node.GetColumnRef()
+		isQualified := false
+		var colName *string
+		// find col name and check if already has schema + table name
+		for _, f := range col.Fields {
+			val := f.GetString_().GetSval()
+			if val == schema {
+				continue
+			}
+			if val == table {
+				isQualified = true
+				break
+			}
+			colName = &val
+		}
+		if !isQualified && colName != nil && *colName != "" {
+			col.Fields = []*pg_query.Node{
+				pg_query.MakeStrNode(schema),
+				pg_query.MakeStrNode(table),
+				pg_query.MakeStrNode(*colName),
+			}
+		}
+	}
+}
+
+func qualifyMysqlWhereColumnNames(sql, schema, table string) (string, error) {
+	stmt, err := sqlparser.Parse(sql)
+	if err != nil {
+		return "", err
+	}
+
+	switch stmt := stmt.(type) { //nolint:gocritic
+	case *sqlparser.Select:
+		err = sqlparser.Walk(func(node sqlparser.SQLNode) (kontinue bool, err error) {
+			switch node := node.(type) { //nolint:gocritic
+			case *sqlparser.ComparisonExpr:
+				if col, ok := node.Left.(*sqlparser.ColName); ok {
+					if col.Qualifier.IsEmpty() {
+						col.Qualifier.Qualifier = sqlparser.NewTableIdent(schema)
+						col.Qualifier.Name = sqlparser.NewTableIdent(table)
+					}
+				}
+				return false, nil
+			}
+			return true, nil
+		}, stmt)
+		if err != nil {
+			return "", err
+		}
+	}
+
+	return sqlparser.String(stmt), nil
 }
